+++ src/test/java/io/loop/ui/UI_step_defs/Polina_UI03_StepDefs.java
@@ -3,15 +3,16 @@ package io.loop.ui.UI_step_defs;
 import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
 import io.loop.ui.pages.POM;
-import io.loop.utilities.BrowserUtils;
-import io.loop.utilities.Driver;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.junit.Assert;
 
-public class DocuportSampleStepDefs {
+/**
+ * Steps created by Polina
+ */
+public class Polina_UI03_StepDefs {
 
-    private static final Logger LOGGER = LogManager.getLogger(DocuportSampleStepDefs.class);
+    private static final Logger LOGGER = LogManager.getLogger(Polina_UI03_StepDefs.class);
     POM pages = new POM();
 
     @When("user inserts {string} to {string} field on {string} page")
@@ -85,34 +86,6 @@ public class DocuportSampleStepDefs {
            default -> throw new IllegalArgumentException("No such a dropdown: " + dropDown);
        }
     }
-    @Then("user should see message {string} on {string} page")
-    public void user_should_see_message_on_page(String massage, String page) {
-       switch (page.toLowerCase().trim()) {
-           case "received doc" -> {
-               Assert.assertEquals(massage, pages.getReceivedDocsPage().getSearchMassage());
-               LOGGER.info("Got " + massage + " massage displayed after Search");
-           }
-           case "invitations" -> {
-               Assert.assertEquals(massage, pages.getInvitationsPage().getSearchMassage());
-               LOGGER.info("Got " + massage + " massage displayed after Search");
-           }
-           default -> throw new IllegalArgumentException("No such a page: " + page);
-       }
-    }
-
-    @Then("user uploads a document")
-    public void user_uploads_a_document() throws Exception {
-//        WebElement element = Driver.getDriver().findElement(By.xpath("//input[@type='file']"));
-//        element.sendKeys("/Users/Polina/Desktop/untitled.txt");
-       // BrowserUtils.uploadFileMac("/Users/Polina/Desktop/untitled.txt");
-        //BrowserUtils.uploadFileUsingAppleScript("/Users/Polina/Desktop/untitled.txt");
-        BrowserUtils.uploadFileMac2("/Users/Polina/Desktop/untitled.txt");
-    }
-
-    @Then("a new tab should open with heading {string}")
-    public void a_new_tab_should_open_with_heading(String termsAndConditionsHeading) {
-       BrowserUtils.switchWindowAndValidate(Driver.getDriver(), "https://numbersquad.com/service-terms-and-conditions", termsAndConditionsHeading);
-    }
 
     @When("user validates {string} is displayed on {string} dropdown")
     public void user_validates_is_displayed_on_dropdown(String text, String dropdown) {
@@ -121,5 +94,13 @@ public class DocuportSampleStepDefs {
             default -> throw new IllegalArgumentException("No such a dropdown: " + dropdown);
         }
     }
+
+
+    @Then("user validates that search results are displayed")
+    public void user_validates_that_search_results_are_displayed() {
+       pages.getReceivedDocsPage().validateSearchResultIsDisplayed();
+       LOGGER.info("Search results are displayed");
+    }
+
 }
